fix: rewrite dispatch and return section imports
