+++ src/app/(main)/dashboard/armamento/despachos/[id]/page.tsx
@@ -7,13 +7,13 @@ import {
   PageHeaderTitle,
 } from '@/modules/layout/templates/page'
 import { buttonVariants } from '@/modules/common/components/button'
-import { ArrowLeft, PackageMinus, PackagePlus } from 'lucide-react'
+import { ArrowLeft, PackageMinus } from 'lucide-react'
 import Link from 'next/link'
 import { getAllProfessionalsToCombobox } from '../../../profesionales/lib/actions/professionals'
 import { getAllReceiversToCombobox } from '../../../armamento/destinatarios/lib/actions/receivers'
 import DispatchesForm from '../../../components/forms/dispatch-form/dispatches-form'
-import { getDispatchById } from '../../../../../../lib/actions/dispatch'
-import { getAllItems } from '../../../../../../lib/actions/item'
+import { getDispatchById } from '@/lib/actions/dispatch'
+import { getAllItems } from '@/lib/actions/item'
 
 export const metadata: Metadata = {
   title: 'Despachos',
