+++ src/app/(main)/dashboard/armamento/despachos/agregar/page.tsx
@@ -11,7 +11,7 @@ import { BackLinkButton } from '@/app/(auth)/components/back-button'
 import { getAllProfessionalsToCombobox } from '../../../profesionales/lib/actions/professionals'
 import { getAllReceiversToCombobox } from '../../../armamento/destinatarios/lib/actions/receivers'
 import DispatchesForm from '../../../components/forms/dispatch-form/dispatches-form'
-import { getAllItems } from '../../../../../../lib/actions/item'
+import { getAllItems } from '@/lib/actions/item'
 
 export const metadata: Metadata = {
   title: 'Despachos',
