+++ src/app/(main)/dashboard/armamento/devoluciones/[id]/page.tsx
@@ -7,14 +7,14 @@ import {
   PageHeaderTitle,
 } from '@/modules/layout/templates/page'
 import { buttonVariants } from '@/modules/common/components/button'
-import { ArrowLeft, IterationCcw, PackagePlus } from 'lucide-react'
+import { ArrowLeft, IterationCcw } from 'lucide-react'
 import Link from 'next/link'
 
 import { getAllProfessionalsToCombobox } from '../../../profesionales/lib/actions/professionals'
 import { getAllReceiversToCombobox } from '../../../armamento/destinatarios/lib/actions/receivers'
-import { getReturnById } from '../../../../../../lib/actions/return'
+import { getReturnById } from '@/lib/actions/return'
 import ReturnsForm from '../../../components/forms/return-form/returns-form'
-import { getAllItems } from '../../../../../../lib/actions/item'
+import { getAllItems } from '@/lib/actions/item'
 
 export const metadata: Metadata = {
   title: 'Devoluciones',
