+++ src/app/(main)/dashboard/armamento/devoluciones/agregar/page.tsx
@@ -7,7 +7,7 @@ import {
   PageHeaderTitle,
 } from '@/modules/layout/templates/page'
 import { getAllItems } from '@/lib/actions/item'
-import { IterationCcw, PackagePlus } from 'lucide-react'
+import { IterationCcw } from 'lucide-react'
 import { BackLinkButton } from '@/app/(auth)/components/back-button'
 import { getAllProfessionalsToCombobox } from '../../../profesionales/lib/actions/professionals'
 import { getAllReceiversToCombobox } from '../../../armamento/destinatarios/lib/actions/receivers'
