+++ src/app/(main)/dashboard/components/forms/return-form/returns-form.tsx
@@ -23,7 +23,7 @@ import es from 'date-fns/locale/es'
 registerLocale('es', es)
 import 'react-datepicker/dist/react-datepicker.css'
 import { ItemsWithAllRelations } from '@/lib/actions/item'
-import { ReturnFormValues } from '../../../abastecimiento/devoluciones/lib/types/types'
+import { ReturnFormValues } from '../../../../../../lib/types/return-types'
 import { FormPeopleFields } from '@/modules/common/components/form-people-fields'
 import { useItemSelector } from '@/lib/hooks/use-item-selector'
 import { ItemSelector } from '@/modules/common/components/item-selector'
